refactor: setupLogRedirect共通化 + Docker read-only設定修正
- setupLogRedirect関数をutils/log-redirect.tsに抽出し、server.ts/watcher.tsの重複を解消
- entrypoint.shでREAD_ONLY=trueを設定し、Docker MCPサーバが意図せずWatcherProcessを起動する問題を修正
