+++ packages/server/src/bin/server.ts
@@ -5,7 +5,7 @@
  */
 
 import * as path from 'path';
-import { readFileSync, mkdirSync } from 'fs';
+import { readFileSync } from 'fs';
 import { FileStorage } from '@search-docs/storage';
 import { DBEngine } from '@search-docs/db-engine';
 import { ConfigLoader, type PidFileContent } from '@search-docs/types';
@@ -16,38 +16,7 @@ import {
   readPidFile,
   isProcessAlive,
 } from '../utils/pid.js';
-import { RotatingWriteStream } from '../utils/rotating-log.js';
-
-/**
- * ログ出力をRotatingWriteStreamにリダイレクト
- */
-function setupLogRedirect(logPath: string): void {
-  mkdirSync(path.dirname(logPath), { recursive: true });
-  const logStream = new RotatingWriteStream(logPath);
-
-  const formatMessage = (args: unknown[]): string => {
-    const timestamp = new Date().toISOString();
-    const message = args.map(a =>
-      typeof a === 'string' ? a : JSON.stringify(a, null, 2)
-    ).join(' ');
-    return `[${timestamp}] ${message}\n`;
-  };
-
-  const originalLog = console.log;
-  const originalError = console.error;
-  const originalWarn = console.warn;
-
-  console.log = (...args: unknown[]) => { logStream.write(formatMessage(args)); };
-  console.error = (...args: unknown[]) => { logStream.write(formatMessage(['[ERROR]', ...args])); };
-  console.warn = (...args: unknown[]) => { logStream.write(formatMessage(['[WARN]', ...args])); };
-
-  // フォアグラウンドモードではコンソールにも出力
-  if (process.stdout.isTTY) {
-    console.log = (...args: unknown[]) => { logStream.write(formatMessage(args)); originalLog(...args); };
-    console.error = (...args: unknown[]) => { logStream.write(formatMessage(['[ERROR]', ...args])); originalError(...args); };
-    console.warn = (...args: unknown[]) => { logStream.write(formatMessage(['[WARN]', ...args])); originalWarn(...args); };
-  }
-}
+import { setupLogRedirect } from '../utils/log-redirect.js';
 
 async function main() {
   try {
